In RDS Utilities (rut) ensure that coord_t always uses 32 bits integers.
* Bug: In <rds>, in rut.h, coord_t is used to read coordinates from gds
    that are on 32 bits. But it was using long that are 64 bits on 64
    bits architectures causing crash. Now uses int32_t.
    Patch contributed by N. Shimizu.

+++ alliance/src/rds/src/rut.h
@@ -13,6 +13,8 @@
 # ifndef RUT_209_H
 # define RUT_209_H
 
+#include <stdint.h>
+
 #ifndef __P
 # if defined(__STDC__) ||  defined(__GNUC__)
 #  define __P(x) x
@@ -163,8 +165,8 @@
 
   typedef struct coord_t
   {
-    long X;
-    long Y;
+    int32_t X;
+    int32_t Y;
 
   } coord_t;
